Allow folder releases from management SDK
