+++ cli/package.json
@@ -39,16 +39,13 @@
     "prompt": "^0.2.14",
     "properties": "^1.2.1",
     "q": "~1.4.1",
-    "recursive-fs": "0.1.4",
     "rimraf": "^2.5.1",
     "semver": "4.3.6",
     "simctl": "0.0.9",
-    "slash": "1.0.0",
     "update-notifier": "^0.5.0",
     "which": "^1.2.7",
     "wordwrap": "1.0.0",
     "xml2js": "^0.4.16",
-    "yargs": "^3.15.0",
-    "yazl": "2.2.2"
+    "yargs": "^3.15.0"
   }
 }
