+++ sdk/definitions/slash.d.ts
