+++ sdk/script/types.ts
@@ -17,4 +17,9 @@ export interface Session {
     machineName: string;
 }
 
+export interface IPackageFile {
+    isTemporary: boolean;
+    path: string;
+}
+
 export type Headers = { [headerName: string]: string };
